Moved tests from lib.rs into a separate file
